requirements.txt
pandas
pyenchant
python-rake
scipy
sklearn
